introduce vouchers.ts with helper functions
